+++ resonate/conventions/local.py
@@ -27,7 +27,7 @@ class Local:
         return None
 
     @property
-    def timeout(self) -> int:
+    def timeout(self) -> float:
         return self.opts.timeout
 
     @property
@@ -40,7 +40,7 @@ class Local:
         idempotency_key: str | Callable[[str], str] | None = None,
         target: str | None = None,
         tags: dict[str, str] | None = None,
-        timeout: int | None = None,
+        timeout: float | None = None,
         version: int | None = None,
     ) -> Local:
         self.id = id or self.id
